\name{NeverConsumersR-package}
\alias{NeverConsumersR-package}
\alias{NeverConsumersR}
\docType{package}
\title{
\packageTitle{NeverConsumersR}
}
\description{
\packageDescription{NeverConsumersR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NeverConsumersR}
\packageIndices{NeverConsumersR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{NeverConsumersR}

Maintainer: \packageMaintainer{NeverConsumersR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
